+++ server/auth-user-service-services/internal/handler/http/handler.go
@@ -55,6 +55,7 @@ func (h *Handler) InitRoutes() *chi.Mux {
 	})
 	r.Route(publicV1, func(r chi.Router) {
 		r.Get("/users/{id}", appMiddleware(h.GetUserByID))
+		r.Delete("/users/{id}", appMiddleware(h.DeleteUserByID))
 	})
 
 	return r
