+++ server/auth-user-service-services/internal/service/user.go
@@ -12,6 +12,7 @@ type IUser interface {
 	CreateUser(ctx context.Context, user entity.User) error
 	GetUserByID(ctx context.Context, id string) (entity.User, error)
 	GetUserByEmailAndPassword(ctx context.Context, email, password string) (entity.User, error)
+	DeleteUserByID(ctx context.Context, id string) error
 }
 
 type User struct {
@@ -24,10 +25,12 @@ func NewUser(client postgres.IUser) IUser {
 	}
 }
 
+// CreateUser - создание пользователя
 func (u *User) CreateUser(ctx context.Context, user entity.User) error {
 	return u.userRepo.CreateUser(ctx, user)
 }
 
+// GetUserByID - получение пользователя по идентификатору
 func (u *User) GetUserByID(ctx context.Context, id string) (entity.User, error) {
 	user, err := u.userRepo.GetUserByID(ctx, id)
 	if err != nil {
@@ -37,6 +40,17 @@ func (u *User) GetUserByID(ctx context.Context, id string) (entity.User, error) 
 	return user, nil
 }
 
+// DeleteUserByID - удаление пользователя по идентификатору
+func (u *User) DeleteUserByID(ctx context.Context, id string) error {
+	err := u.userRepo.DeleteUserByID(ctx, id)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// GetUserByEmailAndPassword - получение пользователя по майлу и паролю
 func (u *User) GetUserByEmailAndPassword(ctx context.Context, email, password string) (entity.User, error) {
 	user, err := u.userRepo.GetUserByEmailAndPassword(ctx, email, password)
 	if err != nil {
